Initial draft

+++ llvm/include/llvm/DebugInfo/GSYM/FileEntry.h
@@ -25,12 +25,12 @@ struct FileEntry {
 
   /// Offsets in the string table.
   /// @{
-  uint32_t Dir = 0;
-  uint32_t Base = 0;
+  uint64_t Dir = 0;
+  uint64_t Base = 0;
   /// @}
 
   FileEntry() = default;
-  FileEntry(uint32_t D, uint32_t B) : Dir(D), Base(B) {}
+  FileEntry(uint64_t D, uint64_t B) : Dir(D), Base(B) {}
 
   // Implement operator== so that FileEntry can be used as key in
   // unordered containers.
@@ -46,16 +46,16 @@ struct FileEntry {
 
 template <> struct DenseMapInfo<gsym::FileEntry> {
   static inline gsym::FileEntry getEmptyKey() {
-    uint32_t key = DenseMapInfo<uint32_t>::getEmptyKey();
+    uint64_t key = DenseMapInfo<uint64_t>::getEmptyKey();
     return gsym::FileEntry(key, key);
   }
   static inline gsym::FileEntry getTombstoneKey() {
-    uint32_t key = DenseMapInfo<uint32_t>::getTombstoneKey();
+    uint64_t key = DenseMapInfo<uint64_t>::getTombstoneKey();
     return gsym::FileEntry(key, key);
   }
   static unsigned getHashValue(const gsym::FileEntry &Val) {
-    return llvm::hash_combine(DenseMapInfo<uint32_t>::getHashValue(Val.Dir),
-                              DenseMapInfo<uint32_t>::getHashValue(Val.Base));
+    return llvm::hash_combine(DenseMapInfo<uint64_t>::getHashValue(Val.Dir),
+                              DenseMapInfo<uint64_t>::getHashValue(Val.Base));
   }
   static bool isEqual(const gsym::FileEntry &LHS, const gsym::FileEntry &RHS) {
     return LHS == RHS;
